+++ packages/fixie-web/CHANGELOG.md
@@ -1,5 +1,12 @@
 # fixie-web
 
+## 1.0.15
+
+### Patch Changes
+
+- Mute audio playback outside of the SPEAKING state
+- Support voice playback transcripts sent as deltas
+
 ## 1.0.14
 
 ### Patch Changes